+++ snippets/page-builder.php
@@ -1,0 +1,7 @@
+<?php
+
+if (isset($field) && $field->isNotEmpty()) {
+    foreach ($field->toBlocks() as $block) {
+        echo $block;
+    }
+}